+++ bin/handler-victorops.rb
@@ -20,7 +20,7 @@ require 'json'
 
 class VictorOps < Sensu::Handler
   option :api_url,
-         description: 'VictorOps API URL without routing key',
+         description: 'VictorOps API URL without routing key. Or use envvar VICTOROPS_API_URL.',
          short: '-a APIURL',
          long: '--api-url APIURL',
          default: nil
@@ -32,12 +32,24 @@ class VictorOps < Sensu::Handler
          default: 'victorops'
 
   option :routing_key,
-         description: 'Routing key',
+         description: 'Routing key. Or use envvar VICTOROPS_ROUTING_KEY.',
          short: '-r KEY',
          long: '--routingkey KEY',
          default: nil
 
+  option :dryrun,
+         description: 'Report settings, and take no action',
+         long: '--dryrun',
+         boolean: true,
+         default: false
+
+
+
   def handle
+    if config[:dryrun]
+      puts "Dryrun: reporting settings and exiting" 
+      puts "  option settingsname set: #{config[:settingsname]}"
+    end  
     # validate that we have a settings name
     unless defined? settings[config[:settingsname]] && !settings[config[:settingsname]].nil?
       raise "victorops.rb sensu setting '#{config[:settingsname]}' not found or empty"
@@ -47,7 +59,23 @@ class VictorOps < Sensu::Handler
     api_url = ENV['VICTOROPS_API_URL']
     api_url = config[:api_url] if api_url.nil?
     api_url = settings[config[:settingsname]]['api_url'] if api_url.nil?
-
+    if config[:dryrun]
+      puts "Determing API_URL to use:"
+      if ENV['VICTOROPS_API_URL']
+        puts "  envvar VICTOROPS_API_URL set: #{ ENV['VICTOROPS_API_URL']}" 
+      else 
+        puts "  envvar VICTOROPS_API_URL not set"
+      end    
+      if config[:api_url]
+        puts "  option api-url set: #{ config[:api_url]}"
+      else
+        puts "  option api-url not set"
+      end  
+      if settings[config[:settingsname]]
+        puts "  settings api_url set: #{ settings[config[:settingsname]]['api_url'] }" if settings[config[:settingsname]]['api_url']
+      end
+      puts "  using: #{api_url}"
+    end
     unless defined? api_url && !api_url.nil?
       raise "victorops.rb sensu setting '#{config[:settingsname]}.api_url' not found or empty"
     end
@@ -56,10 +84,28 @@ class VictorOps < Sensu::Handler
     routing_key = ENV['VICTOROPS_ROUTING_KEY']
     routing_key = config[:routing_key] if routing_key.nil?
     routing_key = settings[config[:settingsname]]['routing_key'] if routing_key.nil?
+    if config[:dryrun]
+      puts "Determing ROUTING_KEY to use:"
+      if ENV['VICTOROPS_ROUTING_KEY']
+        puts "  envvar VICTOROPS_ROUTING_KEY set: #{ ENV['VICTOROPS_ROUTING_KEY']}" 
+      else
+        puts "  envvar VICTOROPS_ROUTING_KEY not set"
+      end
+      if config[:routing_key]
+        puts "  option routingkey set: #{ config[:routing_key]}" 
+      else
+        puts "  option routingkey not set"
+      end
+      if settings[config[:settingsname]]
+        puts "  settings routing_key set: #{ settings[config[:settingsname]]['routing_key'] }" if settings[config[:settingsname]]['routing_key']
+      end
+      puts "  using: #{routing_key}"
+    end
 
     unless defined? routing_key && !routing_key.nil?
       raise 'routing key not defined, should be in Sensu settings or passed via command arguments'
     end
+    return if config[:dryrun]
 
     incident_key = @event['client']['name'] + '/' + @event['check']['name']