+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format located [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Added
+- Adds dryrun option to help diagnose settings being used.
+- Extends usages information to include envvar note
+- Update dev dependancies
 
 ## [2.1.0] - 2019-10-25
 ### Added